# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "OmPay API",
        "description": "API de gestion bancaire avec authentification Passport",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://khadidiatou-fall-api-laravel-0luq.onrender.com",
            "description": "Serveur de production"
        },
        {
            "url": "http://127.0.0.1:8000",
            "description": "Serveur de développement"
        }
    ],
    "paths": {
        "/api/v1/verify-otp": {
            "post": {
                "tags": [
                    "Activation"
                ],
                "summary": "Vérifier le code OTP et activer le compte",
                "description": "Vérifier et activer un compte avec OTP",
                "operationId": "b0e58c338829036fd635253a3c78b3bf",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "telephone",
                                    "otp_code"
                                ],
                                "properties": {
                                    "telephone": {
                                        "type": "string",
                                        "example": "+221771234567"
                                    },
                                    "otp_code": {
                                        "type": "string",
                                        "example": "123456"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Compte activé avec succès et connexion automatique",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Compte activé avec succès. Connexion automatique effectuée."
                                        },
                                        "data": {
                                            "properties": {
                                                "access_token": {
                                                    "type": "string",
                                                    "example": "eyJ0eXAiOiJKV1QiLCJhbGciOiJSUzI1NiJ9..."
                                                },
                                                "token_type": {
                                                    "type": "string",
                                                    "example": "Bearer"
                                                },
                                                "expires_in": {
                                                    "type": "integer",
                                                    "example": 3600
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Données invalides"
                    },
                    "404": {
                        "description": "Token OTP invalide"
                    },
                    "410": {
                        "description": "Code OTP expiré"
                    },
                    "401": {
                        "description": "Code OTP incorrect"
                    }
                }
            }
        },
        "/api/v1/resend-otp": {
            "post": {
                "tags": [
                    "Activation"
                ],
                "summary": "Renvoyer un code OTP",
                "description": "Renvoyer un nouveau code OTP",
                "operationId": "e94eb19a534e3f767b06ecc9b47dc2bb",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "example": "client@example.com"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Nouveau code OTP envoyé",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Nouveau code OTP envoyé"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/modifier-mdp": {
            "post": {
                "tags": [
                    "Activation"
                ],
                "summary": "Modifier le mot de passe PIN",
                "description": "Modifier le mot de passe PIN",
                "operationId": "d2a3c7d27eeefbbfbc84c799dc9242c4",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "ancien_pin",
                                    "nouveau_pin"
                                ],
                                "properties": {
                                    "ancien_pin": {
                                        "type": "string",
                                        "example": "0000"
                                    },
                                    "nouveau_pin": {
                                        "type": "string",
                                        "example": "1234"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Mot de passe modifié avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Mot de passe modifié avec succès"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Données invalides"
                    },
                    "401": {
                        "description": "Ancien PIN incorrect"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT",
                "description": "Enter token in format (Bearer <token>)"
            }
        }
    },
...
    "security": [
        {
            "bearerAuth": []
        }
    ]
}